+++ blockstore/validating_blockstore.go
@@ -12,6 +12,12 @@ type ValidatingBlockstore struct {
 	Blockstore
 }
 
+var _ allKeysChanWithErrer = (*ValidatingBlockstore)(nil)
+
+func (bs *ValidatingBlockstore) allKeysChanWithErr(ctx context.Context) (<-chan cid.Cid, func() error, error) {
+	return allKeysChanWithErrFor(ctx, bs.Blockstore)
+}
+
 func (bs *ValidatingBlockstore) Get(ctx context.Context, c cid.Cid) (blocks.Block, error) {
 	block, err := bs.Blockstore.Get(ctx, c)
 	if err != nil {